Rename source files and directories
- Use lowercase for top-level module name
- Add submodules corresponding to CDA endpoint groups
