+++ cwms/_constants.py
