+++ cwms/exceptions.py
