+++ cwms/timeseries/timeseries.py
@@ -1,7 +1,7 @@
-from .utils import queryCDA, return_df
-import CWMS._constants as constants
-from .core import CwmsApiSession
-from .core import _CwmsBase
+from cwms.utils import queryCDA, return_df
+import cwms._constants as constants
+from cwms.core import CwmsApiSession
+from cwms.core import _CwmsBase
 import pandas as pd
 import json
 import datetime
